pyproject.toml
[tool.poetry]
name = "ragas-eval"
version = "0.1.0"
description = ""
authors = ["Shuichi Kataoka <shu01k9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.8"
openai = "^1.55.0"
faiss-cpu = "^1.9.0.post1"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.9"
langchain-community = "^0.3.8"
ragas = "^0.2.6"
datasets = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
